lib: consolidate TcpRingBuffer into generic RingBuffer with memcpy bulk ops
Remove the duplicate TcpRingBuffer (90 lines) from tcp.rs and extend
the shared RingBuffer<T, N> in lib/ to cover byte-stream use cases:

- Add RingBuffer<u8, N> specialization with write/read/peek_at using
  split copy_from_slice (compiles to memcpy) instead of byte-by-byte
- Modernise indices from u32 to usize, add Clone/Copy derives
- Move read-only queries (len, capacity, is_empty, free_space, peek)
  to the T: Copy impl block so they no longer require Default
- Add generic free_space() and consume() methods

+++ drivers/src/input_event.rs
@@ -372,7 +372,7 @@ pub fn input_has_events(task_id: u32) -> bool {
 pub fn input_event_count(task_id: u32) -> u32 {
     let mgr = INPUT_MANAGER.lock();
     if let Some(idx) = mgr.find_queue(task_id) {
-        mgr.queues[idx].events.len()
+        mgr.queues[idx].events.len() as u32
     } else {
         0
     }
